# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Operations Platform API
  version: 0.1.0
  description: Organization-scoped API contract for currently implemented platform capabilities.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/experiments:
    get:
      operationId: listExperiments
      responses: { '200': { description: Experiments, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Experiment' } } } } } } } }
    post:
      operationId: createExperiment
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateExperiment' } } } }
      responses: { '201': { description: Draft experiment, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '400': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /api/v1/experiments/{experimentID}:
    get:
      operationId: getExperiment
      parameters: [ { $ref: '#/components/parameters/ExperimentID' } ]
      responses: { '200': { description: Experiment, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '404': { $ref: '#/components/responses/Error' } }
  /api/v1/experiments/{experimentID}/approve:
    post: { operationId: approveExperiment, parameters: [ { $ref: '#/components/parameters/ExperimentID' } ], responses: { '200': { description: Approved, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/experiments/{experimentID}/start:
    post: { operationId: startExperiment, parameters: [ { $ref: '#/components/parameters/ExperimentID' } ], responses: { '200': { description: Started or resumed, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/experiments/{experimentID}/pause:
    post: { operationId: pauseExperiment, parameters: [ { $ref: '#/components/parameters/ExperimentID' } ], responses: { '200': { description: Paused, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/experiments/{experimentID}/rollback:
    post:
      operationId: rollbackExperiment
      parameters: [ { $ref: '#/components/parameters/ExperimentID' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string } } } } } }
      responses: { '200': { description: Rolled back, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/experiments/{experimentID}/complete:
    post:
      operationId: completeExperiment
      parameters: [ { $ref: '#/components/parameters/ExperimentID' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExperimentMetrics' } } } }
      responses: { '200': { description: Completed or automatically rolled back, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/experiments/{experimentID}/verify:
    post: { operationId: verifyExperiment, parameters: [ { $ref: '#/components/parameters/ExperimentID' } ], responses: { '200': { description: Verified savings, content: { application/json: { schema: { $ref: '#/components/schemas/Experiment' } } } }, '409': { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: access_token }
  parameters:
...
    ExperimentID: { name: experimentID, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    Error: { description: API error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { type: object, required: [code, message, request_id], properties: { code: { type: string }, message: { type: string }, request_id: { type: string } } } }
...
    ExperimentGuardrails: { type: object, required: [minimum_quality_score, maximum_latency_ms, maximum_error_rate, maximum_cost_per_call], properties: { minimum_quality_score: { type: string }, maximum_latency_ms: { type: string }, maximum_error_rate: { type: string }, maximum_cost_per_call: { type: string } } }
    ExperimentMetrics: { type: object, required: [quality_score, average_latency_ms, error_rate, cost_per_call, control_total_cost, candidate_total_cost], properties: { quality_score: { type: string }, average_latency_ms: { type: string }, error_rate: { type: string }, cost_per_call: { type: string }, control_total_cost: { type: string }, candidate_total_cost: { type: string } } }
    CreateExperiment: { type: object, required: [workload_id, recommendation_id, traffic_percentage, currency, control_execution, candidate_execution, guardrails], properties: { workload_id: { type: string, format: uuid }, recommendation_id: { type: string, format: uuid }, traffic_percentage: { type: string }, currency: { type: string }, control_execution: { type: object, additionalProperties: true }, candidate_execution: { type: object, additionalProperties: true }, guardrails: { $ref: '#/components/schemas/ExperimentGuardrails' } } }
    Experiment:
      type: object
      additionalProperties: true
      required: [id, workload_id, recommendation_id, status, currency, rollback_reason, control_execution, candidate_execution, results, traffic_percentage, guardrails, version, created_at, updated_at]
      properties: { id: { type: string, format: uuid }, workload_id: { type: string, format: uuid }, recommendation_id: { type: string, format: uuid }, status: { type: string, enum: [draft, approved, running, paused, rolled_back, completed, verified] }, currency: { type: string }, rollback_reason: { type: string }, control_execution: { type: object, additionalProperties: true }, candidate_execution: { type: object, additionalProperties: true }, results: { type: object, additionalProperties: true }, traffic_percentage: { type: [string, 'null'] }, verified_savings: { type: [string, 'null'] }, guardrails: { $ref: '#/components/schemas/ExperimentGuardrails' }, started_at: { type: [string, 'null'], format: date-time }, completed_at: { type: [string, 'null'], format: date-time }, version: { type: integer }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } }
security: [ { cookieSession: [] } ]